Isolate process_class_file function to a separate file
